IUserInput: promote promptReal to double values

+++ plugins/org.eclipse.epsilon.eol.dt/src/org/eclipse/epsilon/eol/dt/userinput/JFaceUserInput.java
@@ -191,10 +191,10 @@ public class JFaceUserInput extends AbstractUserInput {
 
 	}
 
-	public float promptReal(String question, float default_) {
-		InputDialog dialog = new InputDialog(Display.getDefault().getActiveShell(),"Real value prompt",question,default_ + "",new FloatValidator());
+	public double promptReal(String question, double default_) {
+		InputDialog dialog = new InputDialog(Display.getDefault().getActiveShell(), "Real value prompt", question, default_ + "", new FloatValidator());
 		if (dialog.open() == Window.OK) {
-			return Float.parseFloat(dialog.getValue());
+			return Double.parseDouble(dialog.getValue());
 		}
 		else {
 			return default_;
